requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==3.1.0
AppiumAndroid==0.1.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver==2.24.1
chromedrivermanager==0.0.1
colorama==0.4.6
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
gevent==23.9.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pip==21.1.2
pluggy==1.3.0
pycparser==2.21
pypiwin32==223
PySocks==1.7.1
pytest==7.4.3
pytest-html==4.1.0
pytest-html-reporter==0.2.9
pytest-metadata==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
requests==2.31.0
robotframework==6.1.1
robotframework-pythonlibcore==4.2.0
robotframework-seleniumlibrary==6.1.3
robotframework-websocketclient==1.3.0
selenium==4.15.1
self==2020.12.3
setuptools==57.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.23.0
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.7
websocket-client==1.6.4
websockets==12.0
wheel==0.36.2
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
